Check if the ADB command succeed before changing the package state
`perform_commands()` now returns a result.

On old devices, ADB commands can return the '0' exit code even if there is an error
On Android 4.4, ADB doesn't check if the package exists. It does not return any error if
you try to `pm block` a non-existent package.

When the command doesn't exist ADB write to stdout "Error: ..." so we can at least catch this.