Automatically retry adb connection until a phone is connected (closes #165)
UAD now can now enter in the `FindingPhones` loading state if it hasn't found a connected device. It can happen at the launch of the app, after pressing the refresh or reboot button.

In this state, the "Finding connected devices..." message is displayed and UAD try every 500ms to initiate an ADB connection with any connected device. If after 1min there is still no device connected then UAD exit the loading state and displays en empty list of packages with the "no devices/emulators found" message.