pkg: Add 'Samsung Services' to list
Adds `com.samsung.android.setupindiaservicestnc` to the list. It shows a
persistent notification after every system update and asks for agreement to
share data with no obvious way to skip. The only way to dismiss it
without agreeing to anything is to click the small text and uncheck all
the items in a list. Then the 'Agree' button becomes a 'Skip' button.