pkg: add `com.samsung.android.setupindiaservicestnc` (#393)
Recommendation list: `Recommended`

Shows a persistent notification after every system update and asks for agreement to share data with no obvious way to skip. The only way to dismiss it without agreeing to anything is to click the small text and uncheck all the items in a list. Then the 'Agree' button becomes a 'Skip' button. Uninstalling the package doesn't have any known side effects.

Co-authored-by: w1nst0n <55300518+0x192@users.noreply.github.com>