refactor:  replaces `drain_filter` (unstable) -with `retain` (stable) (#458)
`drain_filter` was the only unstable feature used in the code. You can now build UAD with Rust Stable.
 
drain_filter is not going to be stabilized anytime soon. The retain method is actually perfectly fine for our usecase. It also operate in place. The benefit of drain_filter was that the inner closure can be mutable. We don't have a need for this.

Co-authored-by: w1nst0n <55300518+0x192@users.noreply.github.com>